create a rule for whole stage codegen
